==> notebooks/41_filter_buildings.ipynb <==
import json
import geopandas as gpd
from collections import defaultdict

file_path = "../data/processed_data/top_500_buildings_with_farms_details.geojson"
with open(file_path, "r") as file:
    geojson_data = json.load(file)

buildings_per_municipality = defaultdict(list)

for feature in geojson_data["features"]:
    municipality_number = feature["properties"]["municipalityNumber"]
    buildings_per_municipality[municipality_number].append(feature)

def total_products(building):
    return sum(len(farm["products"]) for farm in building["properties"]["nearby_farms"])

selected_buildings = []
for municipality, buildings in buildings_per_municipality.items():
    best_building = max(buildings, key=total_products)
    selected_buildings.append(best_building)

output_geojson = {"type": "FeatureCollection", "features": selected_buildings}

output_file_path = "../data/processed_data/top_building_per_municipality.geojson"
with open(output_file_path, "w") as outfile:
    json.dump(output_geojson, outfile, ensure_ascii=False, indent=4)

original_gdf = gpd.read_file("../data/processed_data/top_500_buildings_with_farms_details.geojson")
print(f"original gdf size: {original_gdf.shape}")

filtered_gdf = gpd.read_file("../data/processed_data/top_building_per_municipality.geojson")
print(f"filtered gdf size: {filtered_gdf.shape}")

filtered_gdf.head()
